Split linked_data.py into sub-controllers
linked_data/ package with:
- __init__.py: shared helpers, DI providers, parameter aliases
- lookups.py: LookupController (sources, features, position, hydrolocation)
- navigation.py: NavigationController (modes, info, flowlines, features)
- basin.py: BasinController (basin + split catchment)

All three share path=/linked-data and before_request=check_format.
103 unit tests pass, types clean.
